* feature-0.0.1/wwj:
  新增部件属性功能
  新增部分接口改动

+++ dao/src/main/java/org/thingsboard/server/dao/hs/service/DictDeviceServiceImpl.java
@@ -1,5 +1,7 @@
 package org.thingsboard.server.dao.hs.service;
 
+import com.google.common.base.Joiner;
+import com.google.common.collect.Lists;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.BeanUtils;
@@ -7,6 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.thingsboard.server.common.data.Device;
+import org.thingsboard.server.common.data.DeviceProfile;
 import org.thingsboard.server.common.data.exception.ThingsboardErrorCode;
 import org.thingsboard.server.common.data.exception.ThingsboardException;
 import org.thingsboard.server.common.data.id.DeviceProfileId;
@@ -19,6 +23,9 @@ import org.thingsboard.server.dao.hs.HSConstants;
 import org.thingsboard.server.dao.hs.dao.*;
 import org.thingsboard.server.dao.hs.entity.po.*;
 import org.thingsboard.server.dao.hs.entity.vo.*;
+import org.thingsboard.server.dao.model.sql.DeviceProfileEntity;
+import org.thingsboard.server.dao.sql.device.DeviceProfileRepository;
+import org.thingsboard.server.dao.sql.device.DeviceRepository;
 
 import javax.persistence.criteria.Predicate;
 import java.util.*;
@@ -38,11 +45,20 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
     // 二方库Service
     ClientService clientService;
 
+    // 设备Repository
+    DeviceRepository deviceRepository;
+
+    // 设备配置Repository
+    DeviceProfileRepository deviceProfileRepository;
+
     // 设备字典Repository
-    DictDeviceRepository deviceRepository;
+    DictDeviceRepository dictDeviceRepository;
 
     // 设备字典部件Repository
     DictDeviceComponentRepository componentRepository;
+
+    // 设备字典部件属性Repository
+    DictDeviceComponentPropertyRepository componentPropertyRepository;
 
     // 设备字典属性Repository
     DictDevicePropertyRepository propertyRepository;
@@ -66,7 +82,7 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
      */
     @Override
     public String getAvailableCode(TenantId tenantId) {
-        return this.getAvailableCode(this.deviceRepository.findAllCodesByTenantId(tenantId.getId()), HSConstants.CODE_PREFIX_DICT_DEVICE);
+        return this.getAvailableCode(this.dictDeviceRepository.findAllCodesByTenantId(tenantId.getId()), HSConstants.CODE_PREFIX_DICT_DEVICE);
     }
 
     /**
@@ -95,7 +111,7 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
             return query.where(predicates.toArray(new Predicate[0])).getRestriction();
         };
 
-        return DaoUtil.toPageData(this.deviceRepository.findAll(specification, DaoUtil.toPageable(pageLink)));
+        return DaoUtil.toPageData(this.dictDeviceRepository.findAll(specification, DaoUtil.toPageable(pageLink)));
     }
 
     /**
@@ -107,8 +123,8 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
      */
     @Override
     public DictDeviceVO getDictDeviceDetail(String id, TenantId tenantId) throws ThingsboardException {
-        var dictDevice = this.deviceRepository.findByTenantIdAndId(tenantId.getId(), toUUID(id)).map(DictDeviceEntity::toData)
-                .orElseThrow(() -> new ThingsboardException("dict device not exist", ThingsboardErrorCode.GENERAL));
+        var dictDevice = this.dictDeviceRepository.findByTenantIdAndId(tenantId.getId(), toUUID(id)).map(DictDeviceEntity::toData)
+                .orElseThrow(() -> new ThingsboardException("设备字典不存在！", ThingsboardErrorCode.GENERAL));
 
         var propertyList = DaoUtil.convertDataList(this.propertyRepository.findAllByDictDeviceId(toUUID(dictDevice.getId())))
                 .stream().map(e -> DictDevicePropertyVO.builder().name(e.getName()).content(e.getContent()).build()).collect(Collectors.toList());
@@ -126,8 +142,16 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
                 }).collect(Collectors.toList());
 
         var pMap = componentVOList.stream().collect(Collectors.groupingBy(e -> Optional.ofNullable(e.getParentId()).orElse(HSConstants.NULL_STR)));
+        var componentPropertyList = DaoUtil.convertDataList(this.componentPropertyRepository.findAllByDictDeviceId(toUUID(dictDevice.getId())));
+        var componentPropertyVOList = componentPropertyList.stream()
+                .map(e->{
+                    DictDeviceComponentPropertyVO vo = new DictDeviceComponentPropertyVO();
+                    BeanUtils.copyProperties(e, vo);
+                    return vo;
+                }).collect(Collectors.toList());
+        var cMap = componentPropertyVOList.stream().collect(Collectors.groupingBy(DictDeviceComponentPropertyVO::getComponentId));
 
-        this.recursionPackageComponent(rList, pMap, HSConstants.NULL_STR);
+        this.recursionPackageComponent(rList, pMap, cMap, HSConstants.NULL_STR);
 
         DictDeviceVO dictDeviceVO = DictDeviceVO.builder()
                 .propertyList(propertyList)
@@ -145,15 +169,35 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
      */
     @Override
     @Transactional
+    @SuppressWarnings("Duplicates")
     public void deleteDictDevice(String id, TenantId tenantId) throws ThingsboardException {
-        DictDevice dictDevice = this.deviceRepository.findByTenantIdAndId(tenantId.getId(), toUUID(id)).map(DictDeviceEntity::toData)
-                .orElseThrow(() -> new ThingsboardException("dict device not exist", ThingsboardErrorCode.GENERAL));
-        this.deviceRepository.deleteById(toUUID(dictDevice.getId()));
+        DictDevice dictDevice = this.dictDeviceRepository.findByTenantIdAndId(tenantId.getId(), toUUID(id)).map(DictDeviceEntity::toData)
+                .orElseThrow(() -> new ThingsboardException("设备字典不存在！", ThingsboardErrorCode.GENERAL));
 
-        this.propertyRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
-        this.componentRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
-        this.groupRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
-        this.groupPropertyRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
+        var deviceList = DaoUtil.convertDataList(this.deviceRepository.findAllByTenantIdAndDictDeviceId(tenantId.getId(), toUUID(id)));
+        var deviceProfileId = this.deviceProfileDictDeviceRepository.findByDictDeviceId(toUUID(id))
+                .map(DeviceProfileDictDeviceEntity::getDeviceProfileId).orElse(null);
+        if (deviceList.isEmpty() && deviceProfileId == null) {
+            this.dictDeviceRepository.deleteById(toUUID(dictDevice.getId()));
+
+            this.propertyRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
+            this.componentRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
+            this.componentPropertyRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
+            this.groupRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
+            this.groupPropertyRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
+        } else {
+            StringBuilder sb = new StringBuilder();
+            if (!deviceList.isEmpty()) {
+                var nameList = deviceList.stream().map(Device::getName).collect(Collectors.toList());
+                sb.append("存在关联的设备：").append(Joiner.on(", ").join(nameList));
+            }
+            if (deviceProfileId != null) {
+                var name = this.deviceProfileRepository.findById(deviceProfileId).map(DeviceProfileEntity::getName).orElse("");
+                sb.append("存在关联的设备配置：").append(name);
+            }
+            throw new ThingsboardException(sb.toString(), ThingsboardErrorCode.GENERAL);
+        }
+
     }
 
     /**
@@ -163,7 +207,7 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
      */
     @Override
     public List<DictDevice> listDictDeviceUnused(TenantId tenantId) {
-        return DaoUtil.convertDataList(this.deviceRepository.findAllDictDeviceUnusedByTenantId(tenantId.getId()));
+        return DaoUtil.convertDataList(this.dictDeviceRepository.findAllDictDeviceUnusedByTenantId(tenantId.getId()));
     }
 
     /**
@@ -174,17 +218,19 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
      */
     @Override
     @Transactional
+    @SuppressWarnings("Duplicates")
     public void updateOrSaveDictDevice(DictDeviceVO dictDeviceVO, TenantId tenantId) throws ThingsboardException {
         DictDevice dictDevice = new DictDevice();
         DictDeviceEntity dictDeviceEntity;
         if (!StringUtils.isBlank(dictDeviceVO.getId())) {
-            dictDevice = this.deviceRepository.findByTenantIdAndId(tenantId.getId(), toUUID(dictDeviceVO.getId())).map(DictDeviceEntity::toData)
-                    .orElseThrow(() -> new ThingsboardException("dict device not exist", ThingsboardErrorCode.GENERAL));
-            ;
+            dictDevice = this.dictDeviceRepository.findByTenantIdAndId(tenantId.getId(), toUUID(dictDeviceVO.getId())).map(DictDeviceEntity::toData)
+                    .orElseThrow(() -> new ThingsboardException("设备字典不存在！", ThingsboardErrorCode.GENERAL));
+
             BeanUtils.copyProperties(dictDeviceVO, dictDevice, "id", "code");
 
             this.propertyRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
             this.componentRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
+            this.componentPropertyRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
             this.groupRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
             this.groupPropertyRepository.deleteByDictDeviceId(toUUID(dictDevice.getId()));
         } else {
@@ -193,7 +239,7 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
         }
 
         dictDeviceEntity = new DictDeviceEntity(dictDevice);
-        this.deviceRepository.save(dictDeviceEntity);
+        this.dictDeviceRepository.save(dictDeviceEntity);
 
         AtomicInteger propertySort = new AtomicInteger();
         var propertyList = dictDeviceVO.getPropertyList().stream()
@@ -226,6 +272,7 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
                         .name(t.getName())
                         .sort(groupPropertySort.get())
                         .title(t.getTitle())
+                        .dictDataId(t.getDictDataId())
                         .content(t.getContent()).build();
             }).collect(Collectors.toList());
             r.addAll(dictDeviceGroupPropertyList);
@@ -254,6 +301,21 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
             sort += 1;
             var dictDeviceComponentEntity = new DictDeviceComponentEntity(dictDeviceComponent);
             this.componentRepository.save(dictDeviceComponentEntity);
+
+            if (componentVO.getPropertyList() !=null && !componentVO.getPropertyList().isEmpty()) {
+                int pSort = 1;
+                List<DictDeviceComponentProperty> propertyList = new ArrayList<>();
+                for (DictDeviceComponentPropertyVO propertyVO :componentVO.getPropertyList()) {
+                    var property = new DictDeviceComponentProperty();
+                    BeanUtils.copyProperties(propertyVO, property);
+                    property.setComponentId(dictDeviceComponentEntity.getId().toString());
+                    property.setSort(pSort);
+                    propertyList.add(property);
+                    pSort += 1;
+                }
+                this.componentPropertyRepository.saveAll(propertyList.stream().map(DictDeviceComponentPropertyEntity::new).collect(Collectors.toList()));
+            }
+
             if (componentVO.getComponentList() == null || componentVO.getComponentList().isEmpty()) {
                 continue;
             }
@@ -270,11 +332,13 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
      */
     public void recursionPackageComponent(List<DictDeviceComponentVO> rList,
                                           Map<String, List<DictDeviceComponentVO>> pMap,
+                                          Map<String, List<DictDeviceComponentPropertyVO>> cMap,
                                           String parentId) {
         if (pMap.get(parentId) != null && !pMap.get(parentId).isEmpty()) {
             pMap.get(parentId).forEach(e -> {
+                e.setPropertyList(cMap.getOrDefault(e.getId(), Lists.newArrayList()));
                 rList.add(e);
-                recursionPackageComponent(e.getComponentList(), pMap, e.getId());
+                recursionPackageComponent(e.getComponentList(), pMap, cMap, e.getId());
             });
         }
     }
@@ -301,7 +365,7 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
             if (groupPropertyMap.containsKey(e.getId())) {
                 groupPropertyVOList = groupPropertyMap.get(e.getId()).stream()
                         .map(g -> DictDeviceGroupPropertyVO.builder()
-                                .id(g.getId()).name(g.getName()).content(g.getContent()).title(g.getTitle()).createdTime(g.getCreatedTime())
+                                .id(g.getId()).name(g.getName()).content(g.getContent()).dictDataId(g.getDictDataId()).title(g.getTitle()).createdTime(g.getCreatedTime())
                                 .build()).collect(Collectors.toList());
             }
             r.add(DictDeviceGroupVO.builder().id(e.getId()).name(e.getName()).groupPropertyList(groupPropertyVOList).build());
@@ -319,7 +383,7 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
         var groupPropertyList = DaoUtil.convertDataList(this.groupPropertyRepository.findAllByDictDeviceId(dictDeviceId));
         return groupPropertyList.stream()
                 .map(g -> DictDeviceGroupPropertyVO.builder()
-                        .id(g.getId()).name(g.getName()).content(g.getContent()).title(g.getTitle()).createdTime(g.getCreatedTime())
+                        .id(g.getId()).name(g.getName()).dictDataId(g.getDictDataId()).content(g.getContent()).title(g.getTitle()).createdTime(g.getCreatedTime())
                         .build()).collect(Collectors.toList());
     }
 
@@ -341,9 +405,8 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
      * @param dictDeviceId 设备字典Id
      */
     @Override
-    public DeviceProfileId getDeviceProfileIdByDictDeviceId(UUID dictDeviceId) {
-        var entity = this.deviceProfileDictDeviceRepository.findByDictDeviceId(dictDeviceId);
-        return entity.map(deviceProfileDictDeviceEntity -> DeviceProfileId.fromString(deviceProfileDictDeviceEntity.getDeviceProfileId().toString())).orElse(null);
+    public UUID getDeviceProfileIdByDictDeviceId(UUID dictDeviceId) {
+        return this.deviceProfileDictDeviceRepository.findByDictDeviceId(dictDeviceId).map(DeviceProfileDictDeviceEntity::getDeviceProfileId).orElse(null);
     }
 
     /**
@@ -387,7 +450,23 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
      */
     @Override
     public List<DictDevice> listAllDictDevice(TenantId tenantId) {
-        return DaoUtil.convertDataList(this.deviceRepository.findAllByTenantId(tenantId.getId()));
+        return DaoUtil.convertDataList(this.dictDeviceRepository.findAllByTenantId(tenantId.getId()));
+    }
+
+    /**
+     * 获得全部设备字典属性(包括部件)-数据字典Id Map
+     *
+     * @param dictDeviceId 设备字典Id
+     */
+    @Override
+    public Map<String, String> mapAllPropertyDictDataId(UUID dictDeviceId) {
+        var propertyList = DaoUtil.convertDataList(this.groupPropertyRepository.findAllByDictDeviceId(dictDeviceId));
+        var componentList = DaoUtil.convertDataList(this.componentPropertyRepository.findAllByDictDeviceId(dictDeviceId));
+        var map = componentList.stream().collect(Collectors.toMap(DictDeviceComponentProperty::getName, DictDeviceComponentProperty::getDictDataId, (a, b) -> a));
+        return propertyList.stream().reduce(map, (r, e) -> {
+            r.put(e.getName(), e.getDictDataId());
+            return r;
+        }, (a, b) -> null);
     }
 
     @Autowired
@@ -401,8 +480,8 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
     }
 
     @Autowired
-    public void setDeviceRepository(DictDeviceRepository deviceRepository) {
-        this.deviceRepository = deviceRepository;
+    public void setDictDeviceRepository(DictDeviceRepository dictDeviceRepository) {
+        this.dictDeviceRepository = dictDeviceRepository;
     }
 
     @Autowired
@@ -426,7 +505,22 @@ public class DictDeviceServiceImpl implements DictDeviceService, CommonService {
     }
 
     @Autowired
+    public void setDeviceRepository(DeviceRepository deviceRepository) {
+        this.deviceRepository = deviceRepository;
+    }
+
+    @Autowired
+    public void setDeviceProfileRepository(DeviceProfileRepository deviceProfileRepository) {
+        this.deviceProfileRepository = deviceProfileRepository;
+    }
+
+    @Autowired
     public void setDeviceProfileDictDeviceRepository(DeviceProfileDictDeviceRepository deviceProfileDictDeviceRepository) {
         this.deviceProfileDictDeviceRepository = deviceProfileDictDeviceRepository;
+    }
+
+    @Autowired
+    public void setComponentPropertyRepository(DictDeviceComponentPropertyRepository componentPropertyRepository) {
+        this.componentPropertyRepository = componentPropertyRepository;
     }
 }
